Fix critical hit chance underflow with high target LUK
The local cri variable was truncated to short, but tstatus->luk is
uint16 and can be subtracted with a multiplier of up to 3, allowing
the value to wrap around short's range and produce an incorrect
(often much higher) crit chance instead of a low one. Widen cri to
int to match the actual type of sstatus->cri (int32).

+++ src/map/battle.c
@@ -5155,7 +5155,7 @@ static struct Damage battle_calc_weapon_attack(struct block_list *src, struct bl
 		skill_id == SN_SHARPSHOOTING || skill_id == MA_SHARPSHOOTING ||
 		skill_id == NJ_KIRIKAGE))
 	{
-		short cri = sstatus->cri;
+		int cri = sstatus->cri;
 		if (sd != NULL) {
 			// Racial crit bonuses are affected by katar's crit bonus.
 			if (battle_config.show_katar_crit_bonus && sd->weapontype == W_KATAR)
